Move InstitutionConfigurationOptionsService under Configuration namespace
